drying up demos and renamed components
