+++ unlock-app/src/components/interface/StaticDemo.js
@@ -1,0 +1,11 @@
+import React from 'react'
+import Paywall from '../../pages/paywall'
+import Demo from './Demo'
+
+/**
+ * This is a component that shows the demo with an injected paywall directly in its DOM, not thru
+ * an iframe. It is used only in storybook
+ */
+export default function StaticDemo() {
+  return <Demo paywall={Paywall} />
+}
